Merged PR 6691: Added extra functionality to CheckCompanySpecificPR
- Added information to ReadMe.md
- Added several example pipelines
- Included small fixes

Also tested, for example, the public pipeline that is part of the examples. The pipeline can be found here: https://dev.azure.com/kpn/Azure%20Documentation/_build?definitionId=906

This pipeline will be removed after this PR is approved.

The pipeline for deploying this inside the organization environment:
https://dev.azure.com/kpn/Azure%20Documentation/_build?definitionId=866

Related work items: #19183